Sphinx documentation for pyauditor
