+++ pyauditor/.readthedocs.yaml
@@ -1,0 +1,17 @@
+# https://docs.readthedocs.io/en/stable/config-file/v2.html#supported-settings
+
+version: 2
+
+sphinx:
+  builder: html
+
+build:
+  os: "ubuntu-20.04"
+  tools:
+    python: "3.9"
+    rust: "1.63"
+
+python:
+  install:
+    - method: pip
+      path: .
