+++ pyauditor/docs/conf.py
@@ -1,0 +1,36 @@
+import sphinx_rtd_theme
+
+# Configuration file for the Sphinx documentation builder.
+#
+# For the full list of built-in configuration values, see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+
+project = 'pyauditor'
+copyright = '2022, Stefan Kroboth'
+author = 'Stefan Kroboth'
+
+# -- General configuration ---------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
+
+extensions = [
+        'sphinx.ext.autodoc',
+        'sphinx_rtd_theme'
+]
+
+templates_path = ['_templates']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+
+
+# Interpret first line of docs as function signature
+autodoc_docstring_signature = True
+autodoc_member_order = 'bysource'
+
+# -- Options for HTML output -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
+
+#  html_theme = 'alabaster'
+html_theme = "sphinx_rtd_theme"
+html_static_path = ['_static']
