+++ pyauditor/docs/index.rst
@@ -1,0 +1,25 @@
+.. pyauditor documentation master file, created by
+   sphinx-quickstart on Thu Aug 11 15:26:35 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to pyauditor's documentation!
+=====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+.. automodule:: pyauditor
+   :members:
+   :undoc-members:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
